% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simple_swap_n}
\alias{simple_swap_n}
\title{simple_swap_n}
\usage{
simple_swap_n(Re_from, Re_to, n, swap_p, seed = -1)
}
\arguments{
\item{Re_from}{list of originating vertices for each edge}

\item{Re_to}{list of terminating vertices for each edge}

\item{n}{number of swaps to perform}

\item{swap_p}{base probability of performing a swap (need swap_p < 1 for aperiodicity)}

\item{seed}{optional seed for random number generators}
}
\description{
Perform repeated checkerboard swaps in an adjacency matrix of a simple directed graph
This swapping produces a draws from the uniform distribution of all directed graphs
with the same row and column sums.
}
