% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{swap_n}
\alias{swap_n}
\title{swap_n}
\usage{
swap_n(Re_from, Re_to, n, Rw, Rz_from, Rz_to, seed = -1)
}
\arguments{
\item{Re_from}{list of originating vertices for each edge}

\item{Re_to}{list of terminating vertices for each edge}

\item{n}{number of swaps to perform}

\item{Rz_from}{list of originating vertices for each structural zero}

\item{Rz_to}{list of terminating vertices for each structural zero}

\item{seed}{optional seed for random number generators}

\item{swap_p}{base probability of performing a swap (need swap_p < 1 for aperiodicity).}
}
\description{
Perform repeated checkerboard swaps in an adjacency matrix of a simple directed graph,
respecting structural zeros and swapping according to weights.
This swapping produces a draws from the non-uniform distribution of all directed graphs
with the same row and column sums, with probabilities determined by the edge weights.
}
